Use government interest rate calculations

+++ src/LoanRepaymentApi/UkStudentLoans/PersonDetails.cs
@@ -4,7 +4,7 @@ using System;
 
 public class PersonDetails
 {
-    public decimal AnnualSalaryBeforeTax { get; set; }
+    public int AnnualSalaryBeforeTax { get; set; }
 
     public DateTimeOffset? BirthDate { get; set; }
 }
